Fix third party notices
