+++ packages/compiler/src/index.ts
@@ -35,7 +35,6 @@ export {
   compilerAssert,
   createDiagnosticCollector,
   defineCodeFix,
-  formatDiagnostic,
   getSourceLocation,
   ignoreDiagnostics,
   logDiagnostics,
@@ -101,6 +100,7 @@ export {
   setTypeSpecNamespace,
 } from "./core/library.js";
 export { resolveLinterDefinition } from "./core/linter.js";
+export { formatDiagnostic } from "./core/logger/console-sink.js";
 export { NodeHost } from "./core/node-host.js";
 export { isNumeric, Numeric } from "./core/numeric.js";
 export type { CompilerOptions } from "./core/options.js";
