+++ packages/compiler/test/parser.test.ts
@@ -1,7 +1,8 @@
 import assert, { deepStrictEqual, ok, strictEqual } from "assert";
 import { describe, it } from "vitest";
 import { CharCode } from "../src/core/charcode.js";
-import { formatDiagnostic, logVerboseTestOutput } from "../src/core/diagnostics.js";
+import { logVerboseTestOutput } from "../src/core/diagnostics.js";
+import { formatDiagnostic } from "../src/core/logger/console-sink.js";
 import { hasParseError, parse, visitChildren } from "../src/core/parser.js";
 import {
   IdentifierNode,
